Move Building and BuildingStats class/type to new file buildings.ts
